+++ cronjob/index.js
@@ -8,6 +8,8 @@ import updateNewestNews from './updateNewestNews';
 import updateNewestImages from './updateNewestImages';
 import importCNYES from './importCNYES';
 import importCnaImages from './importCnaImages';
+import importCnaNews from './importCnaNews';
+
 
 
 module.exports = async () => {
@@ -22,6 +24,7 @@ module.exports = async () => {
         updateNewestNews.start(),
         updateNewestImages.start(),
         importCNYES.start(),
-        importCnaImages.start()
+        importCnaImages.start(),
+        importCnaNews.start()
     ]);
 };