+++ CHANGELOG.md
@@ -4,6 +4,7 @@ NOWnews-api change logs
 
 ## [Unreleased]
 ### Fixed
+- api-web 的 /instant 移除匯入新聞（即時跑馬燈）
 - 修正 google trace engine 造成 CPU 與 Memory 消耗過大問題 @SimonSun
 
 ### Added